Refactor analysisd tests
